Breaking up PlatformInput and fixing a focus related crash
